requirements.txt
torch>=2.2.0
transformers>=4.40.0
peft>=0.10.0
pillow>=10.0.0
numpy>=1.26.0
pandas>=2.2.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
seaborn>=0.13.0
jupyter>=1.0.0
accelerate>=0.28.0
bitsandbytes>=0.43.0
einops>=0.7.0
timm>=0.9.0
